eslint fix, security checks while executing commands

+++ commands/misc/say.js
@@ -8,16 +8,23 @@
 /**
  * @type {import('../../typings').LegacyCommand}
  */
+const { PermissionsBitField } = require("discord.js");
 module.exports = {
 	name: "say",
 
-	execute(message, args) {
+	async execute(message, args) {
 		try {
-			let text = args.join(" ");
-			message.delete();
-			message.channel.send(text);
+			const text = args.join(" ");
+			if (
+				message.guild.members.me.permissions.has(
+					PermissionsBitField.Flags.ManageMessages
+				)
+			) {
+				message.delete();
+				message.channel.send(text);
+			}
 		} catch (error) {
-			console.log(error.msg);
+			console.log(error);
 		}
 	},
 };
